WebService/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
gcloud==0.17.0
googleapis-common-protos==1.5.5
httplib2==0.12.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jws==0.1.3
MarkupSafe==1.1.0
numpy==1.15.4
oauth2client==3.0.0
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scipy==1.1.0
six==1.11.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
